Работа над вторым домашним заданием, переделана часть классов и взаимодействий для соответствия шаблону "Композит"
