+++ src/main/java/ru/sbt/mipt/oop/CollectionSmarthome/Room.java
@@ -1,11 +1,13 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.CollectionSmarthome;
+
+import ru.sbt.mipt.oop.EventProcessing.SensorEvent;
+import ru.sbt.mipt.oop.PartofTheHouse;
 
 import java.util.Collection;
 
 public class Room extends PartofTheHouse implements ComponentOfTheCollection {
     private Collection<Light> lights;
     private Collection<Door> doors;
-    private String name;
 
     public Room(Collection<Light> lights, Collection<Door> doors, String name) {
         this.lights = lights;
@@ -27,10 +29,10 @@ public class Room extends PartofTheHouse implements ComponentOfTheCollection {
 
     @Override
     public void execute(SensorEvent event, PartofTheHouse room) {
-        for (Light light: getLights()) {
+        for (Light light : getLights()) {
             light.execute(event, this);
         }
-        for (Door door: getDoors()) {
+        for (Door door : getDoors()) {
             door.execute(event, this);
         }
     }
