--- src/main/java/ru/sbt/mipt/oop/ComponentOfTheCollection.java
@@ -1,5 +1,0 @@
-package ru.sbt.mipt.oop;
-
-public interface ComponentOfTheCollection {
-    public void execute(SensorEvent event, PartofTheHouse part);
-}
