--- src/main/java/ru/sbt/mipt/oop/EventHandler.java
@@ -1,5 +1,0 @@
-package ru.sbt.mipt.oop;
-
-public interface EventHandler {
-    public void handleEvent(Room room, SensorEvent event, SmartHome smartHome);
-}
