+++ src/main/java/ru/sbt/mipt/oop/EventProcessing/DoorEventHandler.java
@@ -1,0 +1,14 @@
+package ru.sbt.mipt.oop.EventProcessing;
+
+import ru.sbt.mipt.oop.CollectionSmarthome.Room;
+import ru.sbt.mipt.oop.CollectionSmarthome.SmartHome;
+
+public class DoorEventHandler implements EventHandler {
+    @Override
+    public void handleEvent(Room room, SensorEvent event, SmartHome smartHome) {
+        smartHome.execute(event, smartHome);
+    }
+
+
+}
+
