+++ src/main/java/ru/sbt/mipt/oop/EventProcessing/EventHandler.java
@@ -1,0 +1,8 @@
+package ru.sbt.mipt.oop.EventProcessing;
+
+import ru.sbt.mipt.oop.CollectionSmarthome.Room;
+import ru.sbt.mipt.oop.CollectionSmarthome.SmartHome;
+
+public interface EventHandler {
+    public void handleEvent(Room room, SensorEvent event, SmartHome smartHome);
+}
