+++ src/main/java/ru/sbt/mipt/oop/EventProcessing/LightEventHandler.java
@@ -1,4 +1,7 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.EventProcessing;
+
+import ru.sbt.mipt.oop.CollectionSmarthome.Room;
+import ru.sbt.mipt.oop.CollectionSmarthome.SmartHome;
 
 public class LightEventHandler implements EventHandler {
     @Override
