+++ src/main/java/ru/sbt/mipt/oop/EventProcessing/SensorEvent.java
@@ -1,12 +1,20 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.EventProcessing;
+
+import ru.sbt.mipt.oop.CollectionSmarthome.SmartHome;
 
 public class SensorEvent {
     private final SensorEventType type;
     private final String objectId;
+    private final SmartHome smartHome;
 
-    public SensorEvent(SensorEventType type, String objectId) {
+    public SensorEvent(SensorEventType type, String objectId, SmartHome smartHome) {
         this.type = type;
         this.objectId = objectId;
+        this.smartHome = smartHome;
+    }
+
+    public  SmartHome getSmartHome (){
+        return smartHome;
     }
 
     public SensorEventType getType() {
