+++ src/main/java/ru/sbt/mipt/oop/EventProcessing/SensorEventType.java
@@ -1,4 +1,4 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.EventProcessing;
 
 public enum SensorEventType {
     LIGHT_ON, LIGHT_OFF, DOOR_OPEN, DOOR_CLOSED
