+++ src/main/java/ru/sbt/mipt/oop/Input/ReadFromFile.java
@@ -1,6 +1,7 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.Input;
 
 import com.google.gson.Gson;
+import ru.sbt.mipt.oop.CollectionSmarthome.SmartHome;
 
 import java.io.IOException;
 import java.nio.file.Files;
