+++ src/main/java/ru/sbt/mipt/oop/PartofTheHouse.java
@@ -1,8 +1,10 @@
 package ru.sbt.mipt.oop;
 
 public class PartofTheHouse {
-    private String name;
+    public String name;
+
     public String getName() {
         return name;
     }
 }
+
