+++ src/main/java/ru/sbt/mipt/oop/Sender.java
@@ -1,7 +1,7 @@
 package ru.sbt.mipt.oop;
 
 public class Sender {
-    static void sendCommand(SensorCommand command) {
+    public static void sendCommand(SensorCommand command) {
         System.out.println("Pretent we're sending command " + command);
     }
 }
