+++ src/main/java/ru/sbt/mipt/oop/SmartHomeScenarios/SmartHomeScenarioProcessor.java
@@ -1,4 +1,6 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.SmartHomeScenarios;
+
+import ru.sbt.mipt.oop.CollectionSmarthome.SmartHome;
 
 public class SmartHomeScenarioProcessor {
