+++ src/main/java/ru/sbt/mipt/oop/SmartHomeScenarios/SmartHomeScenarioTypes.java
@@ -1,4 +1,4 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.SmartHomeScenarios;
 
 public enum SmartHomeScenarioTypes {
     HALL_DOOR_WAS_CLOSED
